requirements.txt
alembic==1.13.2
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
asyncpg==0.29.0
autoflake==2.3.1
bcrypt==4.2.0
billiard==4.2.0
black==24.8.0
celery==5.4.0
certifi==2024.8.30
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.112.2
fastapi-cache2==0.2.2
fastapi-slim==0.112.2
fastapi-versioning==0.10.0
flake8==7.1.1
flower==2.0.1
greenlet==3.0.3
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
humanize==4.10.0
idna==3.8
iniconfig==2.0.0
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
kombu==5.4.0
Mako==1.3.5
MarkupSafe==2.1.5
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.1
passlib==1.7.4
pathspec==0.12.1
pendulum==3.0.0
pillow==10.4.0
platformdirs==4.3.2
pluggy==1.5.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
psycopg2==2.9.9
pyasn1==0.6.0
pycodestyle==2.12.1
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic_core==2.20.1
pyflakes==3.2.0
pyright==1.1.379
pytest==8.3.2
pytest-asyncio==0.24.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-json-logger==2.0.7
python-multipart==0.0.9
pytz==2024.1
redis==4.6.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
sqladmin==0.18.0
SQLAlchemy==2.0.32
starlette==0.38.2
time-machine==2.15.0
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.1
uvicorn==0.30.6
vine==5.1.0
wcwidth==0.2.13
WTForms==3.1.2
